Split InputMgr into it's own header file.
Mainly so I can split view it..
